start spitting up
